Refactored folder structure, renamed test files, main is now executable
