+++ src/__init__.py
@@ -1,1 +1,3 @@
-# This file is intentionally left blank.
+# This file marks the 'src' directory as a Python package.
+# It allows you to import modules from this folder using package-style imports.
+# You can leave this file empty unless you want to run initialization code for the package.
