+++ tests/test_CmdApp.py
@@ -1,10 +1,10 @@
 import sys
 import os
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../src')))
 
 import pytest
 from unittest.mock import patch, Mock
-from cmdApp import CMDApp
+from cmd_app import CMDApp
 
 @pytest.fixture
 def app():
